Change file name and move fuinction to another file
